add SIGTERM handling

+++ libs/general/killBot.js
@@ -4,9 +4,8 @@ const { Client } = require('discord.js');
  * @param {Client} client
  */
 function killBot(client) {
-  if (!client || !(client instanceof Client)) return;
-
-  client.destroy();
+  if (client && ('destroy' in client)) client.destroy();
+  console.debug('Bye!');
   process.exit(0);
 }
